shill: Perform synchronous Get before reading property value
Currently, when getting a property value, we only read the
cached value. It is possible to read the cache value before
it is initialized, since it is being initialized using
asynchronous GetAll function.

Instead, explicity perform synchronous Get to ensure the
cache value is valid before returning it.  This will match
the previous behavior that we had when using dbus-c++ library.

This will fix the problem with WiMax, where we tried to
use an uninitialize property value, causing WiMax service
failed to start.

Bug: None
BUG=chromium:533423
TEST=Run network_WiMaxSmoke autotest
TEST=Run wifi_matfunc

Change-Id: I0bef30a45e5ab34ffc142e6076755f9d74107af9