shill: Defer control interface destruction
D-Bus control interface owns a bus connection for proxy objects.
While stopping the Manager, devices such as WiMax will perform
asychronous method calls (Disable) over the D-Bus. The async method
call will be deferred and posted to the event loop with a refptr to
the ObjectProxy.

Shutting down the proxy bus before this async method call is ran will
not destroy the ObjectProxy for the WiMax device. It will get destroyed
when the async method call is ran, which causes fatal error in its
destructor since there is a pending call:
[FATAL:object_proxy.cc(63)] Check failed: pending_calls_.empty()

So defer the destruction of the control interface |control_| to
when the daemon is destroyed. Since it is guarantee that there
will not be any task left on the message loop by that time.
This ensures that all ObjectProxy will be correctly destroyed
when bus is shutdown (Bus::ShutdownAndBlock).

Bug: None
BUG=chromium:537771
TEST=Restart shill on device with WiMax doesn't crash

Change-Id: Ib11ddf268b505098f29f8f1c6fd3034c038b9925