quipper: Pass perf_event_attr to SampleInfoReader
SampleInfoReader depends on multiple members of struct perf_event_attr.
e.g. sample_type, read_info, and in the future sample_{regs|stack}_user.
It would be impractical to keep adding parameters for individual fields,
so pass in the entire struct.

The read/write size checking is now based around the size of the entire
event_t struct, and not just the dangling sample info.

BUG=chromium:537800
TEST=unit tests pass

Change-Id: I2e20a654ce846edb6163af407488504f1013ec8f
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/303138
Reviewed-by: David Sharp <dhsharp@chromium.org>