shill: test-proxy: Create a shill client interface (1/3).
Creating base shill client abstract class which can be used to interface
with Shill over dbus, binder, etc.

ProxyShillWifiClient: This is the abstract class which can be
implemented to interface with Shill for wifi tests. The RPC server will
invoke methods defined in this interface. This layering will help us
migrate to a different IPC mechanism in the future (binder instead of
dbus).

Bug: 24335496
Change-Id: I56d2de49fe5d6887b891a0ffd51082e6961fdc2e
TEST: mmm system/connectivity/shill/test-rpc-proxy