init: stop failsafe-delay after failsafe is started
We want to stop once failsafe is started.  If we don't do that then we
can end up re-trigging failsafe (oops) if shutdown happens within the
first 30 seconds after boot-services because:
- failsafe will stop when system-services stops
- failsafe delay may finish its 30 second sleep after failsafe stops
- failsafe will start again

Note that on mickey this could be particularly bad because we'd be
likely to trigger the failsafe delay after we woke up from s3halt, then
start trying to trigger udev!

BUG=chromium:538330
TEST=gzip ui.conf; reboot; see that open-ssh still starts
TEST=reboot, login quickly, see that failsafe-delay is not running

Change-Id: Ibedd0c25c1dec179eb148146295d82c23816b9a6
Reviewed-on: https://chromium-review.googlesource.com/303535
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>