shill: Do not attempt to connect if supplicant is not present
When shill and apmanager is used by weaved for WiFi bootstrapping,
weaved will instruct shill to connect to a WiFi network right
after apmanager release the device back to shill. shill will
recreate the WiFi device after getting the device back. In this
case, the connect will happen before supplicant is detected by
shill.  This causes shill to crash since it will attempt to use
supplicant proxy to initiate the connection.

Fix this by ignoring the connection attempt if supplicant is not
present.  Once supplicant is detected, shill will auto-connect to
the target network (since weaved will configure it as auto-connectable)
when it is discovered in the scan result.

Bug: 24605760
TEST=Setup device using weaved, verify shill doesn't crash from logcat.
TEST=Run unittest on Chrome OS.

Change-Id: I3a61ac00bb12c8f2350db9aff1dca66e8bbb9070