minijail: fix input validation in gen_constants.sh
Add missing whitespace to input validation check in
gen_constants.sh, so that this check actually runs.

While there, reduce the 3-argument version of this
script to 2 arguments by combining the path to the
compiler binary and the arguments to pass to
the compiler into one argument.

BUG=chromium:539662
TEST=minijail builds using emerge.

Change-Id: Ib31688b034d960b6dbb2cb241401ef560f3c517f
Reviewed-on: https://chromium-review.googlesource.com/304116
Commit-Ready: Samuel Tan <samueltan@chromium.org>
Tested-by: Samuel Tan <samueltan@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>