shill: test-proxy: Create a module for utility functions.
Create a free standing module for utility functions.
Implement the XmlRpcValue to BrilloAny conversions needed in
ProxyRpcServer to plumb the values recieved from the external xmlrpcpp
library to dbus values and vice versa. These are specifically needed for
|get_service_properties|, |set/get_dbus_property_on_device| RPC methods
which will need to convert to/from brillo::Any values to the appropriate
type for that property.

Bug: 25443143
Change-Id: I9124c118fd92059a9bb71df95f5c0245ff633a12
Test: mmm system/connectivity/shill/test-rpc-proxy