shill: don't truncate sequence numbers
Both when handling NetlinkPackets and when managing the pending dump
sequence number shill has previously truncated the 32bit netlink
sequence to a uint16_t.  This causes issues where the untruncated
sequence is used as the key in the map of handlers, meaning that the
handler never gets invoked and the messages are then treated as if they
were broadcast messages.

Bug: None
BUG=chrome-os-partner:46556
TEST=created a program that issued more than 2^16-1 dump requests

Change-Id: I2d68cd61d5de4cefd50a1f229e4dc5f08cab77ee