shill: test-proxy: Implement wait for a property state.
Create a generic templated function to wait for a property to
change to one of the specified set of values on a given proxy.
Uses the message loop to block the calling function until the Dbus
signal callback indicates that the property value is in
one of the expected states, or until a timeout occurs.

Related Changes:
a. Create a generic |WaitForEvent| function to wait for any event to
occur within the provided timeout.
b. Move the Dbus callback functions to free standing ones from
member functions.
c. Remove the |ComparePropertyValueIn| function and use
|GetProxyPropertyValue| instead.

Bug: 24335496
Change-Id: Ia36d7226bc91a2a1041916ff557ff61beacf4baa
Test: mmm system/connectivity/shill/test-rpc-proxy