shill: test-proxy: Implement wait for a given service.
Add a helper function to wait for a service proxy matching the specified
properties. Uses the message loop to block the calling function until
the specified service is found, or if timeout occurs.
The caller specifies a rescan interval which is used by the
function to schedule periodic checks for the existence of the specified
service. The helper function triggers a rescan of the specified
technology if it doesn't find the service after each rescan interval.

Bug: 24335496
Change-Id: I30a43869a9295de176455b9d3d587d032f91a5ae
Test: mmm system/connectivity/shill/test-rpc-proxy