shill: Re-enumerate renamed devices if they were blacklisted
In general, shill ignores name changes on devices.  For
Chrome OS systems renames never happen unless done in dev
mode from the command line.  In other systems than Chrome OS
where shill is used there may be udev rules or other actors
that rename devices while shill is running.

In the case where shill creates a Device instance and is
actively using the device while it is renamed, this causes
a number of problems since the Device instance has changed
names and a number of operations shill performs on the
device will continue to use the defunct name.  This CL
does not address this scenario.

Instead, this CL addresses a more easy-to-handle case
where shill first discovers a blacklisted device and
hence does not actively use it.  If such a device is
renamed, shill will now act as if the device has been
newly added, possibly undoing the blacklist.  This allows
use cases where the role of a device when it appears is
not known ahead of time and some system service will
assign it to a known role later.

To implement this, the code for retrieving a link name
from an RTNL message is factored out and used in a new
helper method that tests whether an incoming RTNL message
indicates a device rename of a previously blacklisted
device.

BUG=chrome-os-partner:42699
TEST: New unit tests + manual:
TEST: Run shill on Edison with "--device-black-list=wlan0" argument
TEST: Run "ip link set wlan0 name wlan0b"
TEST: Observe that shill ignored wlan0 but initialized and used wlan0b

Change-Id: I70975d55292d4f4f330b33ec2d0439a568148588