apmanager: Use asynchronous handler for org.chromium.Service.Start D-Bus call
This allows us to handle this D-Bus call asynchronously for Brillo
devices.  Since we want to wait for the newly created AP interface
to be enumerated before starting the service.

For now, this handler will return synchronously on both Chrome OS
and Brillo.  A separate CL (CL:178864) will update this handler to
return asynchronously on Brillo.

While there, remove the unnecessary mock function for Service.Start
and Service.Stop, since these are D-Bus method handlers.  Even though
the MockService class is now empty, we might still need it when we
remove the D-Bus dependency from the Service class (b/24194427),
so keep it for now.

BUG=b:25113165
TEST=Unittest on both Chrome OS and Brillo
TEST=Manual test on both Chrome OS and Brillo

Change-Id: If0ba13ae89a25bdf9207803d7d6b008e9ad6165e