shill: NetlinkAttribute: Accept multiple int sizes
The RegInitiator parameter in CRDA-related nl80211 messages can
be encoded with two different data sizes depending on context.
In some contexts it is a u32 and in others it is a u8.  This CL
allows shill to accept this parameter in either of those
encodings, instead of erroring out on a message with the u8
encoding.

While here, make "Invalid |input|" messages more verbose about
the type they are trying to decode and the size of the data
that failed to decode.

BUG=chromium:549611
TEST:Unit test

Change-Id: I62eafc9ca8d708826275c4a2e97f1b2f62964e36