login: thaw renderers before killing processes
Chrome renderers are frozen in some circumstances, such as during
dark resume. If Chrome crashes during this time, the renderers will
be left in the freezer, and we won't be able to kill them when we
try to restart the UI. This leaves renderers orphaned.

BUG=chromium:551116
TEST=run as root:
  # echo FROZEN > \
    /sys/fs/cgroup/freezer/chrome_renderers/to_be_frozen/freezer.state
  # ps aux | grep renderer
  # restart ui
  # ps aux | grep renderer
  and make sure the first set of renderers is not hanging around.

Change-Id: I04ec026472b85d4bcd841e6bf9b645996156bf2a
Reviewed-on: https://chromium-review.googlesource.com/311309
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>