quipper: PerfOptionParser: Fix a past-the-end bug for value arguments
If a value argument is at the end of the vector, args_iter would get
incremented twice. Since the loop condition is an equality comparison with
the end iterator, the for loop could go beyond the bounds of the vector.

Check for being at the end after skipping over an argument value, and add
some test cases.

TEST=Added unit test
BUG=chromium:537464

Change-Id: I3eca02641e51ed52fc0fee50de1f0fd1a7adb6e3
Reviewed-on: https://chromium-review.googlesource.com/311834
Commit-Ready: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>
Reviewed-by: Emma Rapati <rapati@chromium.org>