apmanager: dbus: perform D-Bus initialization in DBusControl
This moves the RPC specific initialization out of the Daemon class,
which makes it agnostic of the underlying RPC implementation.

Also move the ownership of Manager from Daemon to the ControlInterface
implementation. So the ControlInterface is now responsible for
creating Manager and exposing it through the underlying RPC interface.
This makes it easier to maintain the relevant objects (e.g. Manager,
ObjectManager), and avoids possible confusion during cleanup.

BUG=b:25654681
TEST=Verify device setup works on Brillo device

Change-Id: Id2eb8d08346841e294f8286fe191708caa2cbc24