shill: ProcessManager: Forget self-reaped processes
In ProcessManager::StopProcessAndBlock(), the process manager
attempts to reap child processes itself, not waiting for
ProcessReaper.  The exit processes reaped in this manner will
go unnoticed by ProcessReaper.  If we do not manually remove
such processes from ProcessReaper, this stale record will get
in the way of a later ProcessReaper::WatchForChild() if this
pid is ever used again (e.g., when the system's pid counter
rolls over).  This will cause a LOG(FATAL) in some situations.

To fix this, this CL adds a call to ProcessReaper::ForgetChild()
whenever ProcessManager::StopProcessAndBlock() is executed.

BUG=chromium:535910
CQ-DEPENDS=183117

Change-Id: I041e97ac799210229928dba1278e7baa0f49bfe0