power: Add ReleaseAllDisplayWakeLocks method
This should be used to clean up the wake lock state when ARC
crashes. ARC will not realize there are still wake locks held as
it has lost all of the state in its power manager and thus we
need to get rid of all of them.

CQ-DEPEND=CL:313295
BUG=None
TEST=unittests for glados, manual testing on minnie with dbus-send

Change-Id: Ieed51610c47a368d86ef65d2142644a1a23358f8
Reviewed-on: https://chromium-review.googlesource.com/313197
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>