shill: check connectivity before calling OnConnectedAndReachable
We currently call WakeOnWiFi::OnConnectedAndReachable whenever our
IPv6 configuration is updated, even though this does not always means
that we have established IPv6 connectivity and IP-reachability.
Since kernel notifications for adding and deleting IPv6 are not
synced with the AP association process, we sometimes update our IPv6
configuration even before we have completed associating with an AP.
When this happens during dark resume, shill mistakenly assumes that
it is connected and pre-maturely re-suspends, preventing the association
process from completing.

Fix this by calling WiFi::IsConnectedToCurrentService() to check if
we are connected, before calling WakeOnWiFi::OnConnectedAndReachable
in WiFi::OnIPv6ConfigUpdated.

BUG=chromium:555722
TEST=shill compiles and all unit tests pass.

Change-Id: I8012886611854f44b7a6fcd4b780e3a5775429b6