p2p: Reduce the size and increase the timeout of HttpServer.Basic.
The HttpServer.Basic tests the interaction between the HttpServer and
its worker threads. The test runs locally on a workstation with no load
in 50 ms, but a overloaded builder can take more than a minute to
schedule a ready process.

This patch reduces the thread number used in the test from 5 to 3, to
reduce the number of context switches required to complete the test and
increases the timeout from 10 to 30 minutes.

BUG=chromium:554682
TEST=FEATURES=test emerge-link p2p

Change-Id: Id18b34c19cc4e090f2b0af68d3528450effb5d1f
Reviewed-on: https://chromium-review.googlesource.com/314002
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: David Zeuthen <zeuthen@chromium.org>