cryptohome: Set attestation database ownership.
By default the file ownership will be root:root but on some platforms
attestationd also needs to read this file and it runs as the system user
'attestation'. This change ensures that the database ownership is set to
and stays as attestation:attestation.

BUG=chrome-os-partner:46244
TEST=manual, unit
CQ-DEPEND=CL:315225

Change-Id: Iaaa8c58655fde674165e6f973ba7a61c81e7e8fa
Reviewed-on: https://chromium-review.googlesource.com/315147
Commit-Ready: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Jeromy Leugers <jleugers@chromium.org>
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>