shill: do not leak opened file descriptors to the child process
When spawning a new process, explicitly ask it to close all unnecessary
opened file descriptors inherited from the parent process.

This used to be taken care of by the glib's process library, by not
setting G_SPAWN_LEAVE_DESCRIPTORS_OPEN flag when spawing a process.
We have to do that manually now that we moved over to used libbrillo's
process library.

Bug: None
BUG=chromium:531655
TEST=Setup a VPN connection, verify ARP socket does not leak to
TEST=the VPN process and no more spammy ARP receive errors.

Change-Id: I019bf0461155313d6726bb793fc81f777d53ae75