crash: Add collector for ARC crashes
Initial CL for ARC crash reporting. If crash_reporter is built with
ARC support and ARC is running, crashes for ARC processes are ignored
by the user collector, and handled by the ARC collector. This collector
uses the ARC package name as the executable name in reports, and checks
if the crash involved a system process before generating a dump. Crash
dump and metadata generation for the ARC collector will be implemented
in subsequent CLs.

BUG=b:25773929
TEST=kill -SEGV <pid>
     grep crash_reporter /var/log/messages
     ls /var/spool/crash  # Reports should use the package name.

Change-Id: I50ab712b504d12a38d0744965d016c95bb212b9b
Reviewed-on: https://chromium-review.googlesource.com/317859
Commit-Ready: Dominik Laskowski <domlaskowski@google.com>
Tested-by: Dominik Laskowski <domlaskowski@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>