shill: process_manager: reset signal mask for minijailed child processes
By default, minijailed child processes will inherit signal mask
(blocked signals) from its parent process.  When a signal is blocked,
the child process will not be able to receive that signal.

To fix it, set a flag to instruct the child process to reset the signal
mask when starting, so that the child process will be started without
any signals being blocked, which allows it to correclty configure its
own signal handling.

Bug: None
BUG=chrome-os-partner:47785
TEST=Verify dhcpcd can be killed using SIGTERM "kill -15 <pid>"

Change-Id: Idb9fce9a02a1459cb8d7ee1f9bdfa611197ef639