shill: Disable RTTI
There is no longer a reliance on RTTI in libbrillo, so disable RTTI
in the rest of Brillo codebase.

Shill also used dynamic_cast<> in many places but it looks like there
is no reason for that. Using static_cast<> in those places seems to
work just as well and doesn't require RTTI.

Bug: 26292405, chromium:220387
TEST: Compiles, all unit tests pass on Dragonboard
      Was able to bootstrap a device from scratch and networking
      seems to work fine.

Change-Id: Id0c47168e634e354b4279f613246e6667df6fb5f