login: freeze and thaw ARC instance across suspend/resume
This allows us greater control of the ARC instance, such as not
letting the processes wake up in dark resume. Until we implement
something more like Doze mode, this is useful to keep the
instance from becoming unhappy when we suspend and resume a lot.

BUG=b:25410226
TEST=deploy to device, check for relevant log messages;
  new unit test

Change-Id: I37193c04bfdd5b5a431af0d2ad2cc6ae809b0269
Reviewed-on: https://chromium-review.googlesource.com/319819
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>