quipper: test_perf_data: Attr id support; Consolidate file header classes
Add support for writing the attr id section. Previously this was always an
empty section, with attrs immediately following the header. The attr ids
section goes after the header, before the attrs. The perf_file_attr.ids
refers to a segment of this section. Add a class (AttrIdsSection) to
write the attr ids, and have it return perf_file_section structs that can
be written to perf_file_attr.ids.

Because of the inserted attr ids section, the offsets for the attrs and
data sections in the (normal mode) file header must now be calculate
rather than using nominal values. And because of this, it is simpler to
move the functionality of class ExamplePerfDataFileHeader_CustomAttrSize
into ExamplePerfDataFileHeader::WithCustomAttrSize(). The updated class
re-calculates the offsets of the various sections upon calling the
With...() methods.

TEST=unit tests
BUG=none

Change-Id: Icf4ff7166ee112ee0dafd587db84bc38b34fdf40
Reviewed-on: https://chromium-review.googlesource.com/320611
Commit-Ready: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>
Reviewed-by: David Sharp <dhsharp@chromium.org>