webserver: libwebserv: Remove unused Disconnect() method.
DBusServer::Disconnect() method was not used anywhere. Instead, the
unique_ptr<Server> can be just destroyed to make it disconnect from
DBus. Note that this matches the connection side, since it is only
connected during construction (via factory in the Server class).

The order of the private data members was changed to match the
destruction order used in the Disconnect() method.

BUG=None
TEST=FEATURES=test emerge-stom webserver

Change-Id: I2552a40bd87c54bf35f22718e9d0230011e7e4f6