Add libcontainer
This library provides a way to start a process in new namespaces and a
new root file system.  It jails the process using libminijail.  This
will be used by session manager in coming commits.  All mounts outside
of the minijail are enforced to be noexec.

User namespacing is disabled and will be added separately in order to
minimize disruption from this change.

BUG=b/26549867
TEST=new unittests and a busybox starting example.

Change-Id: I4428f4b173a07c75b0b584dba87da0ede1f66f94
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/323433