debugd: Fix the grep pattern for touch_fw_version
It looks like the old grep pattern hasn't been valid for a while here,
and wasn't actually finding any touch FW updating logs when you went
to chrome://system.  This CL alters the grep expression slightly so
that it should match the new format.

BUG=chromium:582632
TEST=built and emerged onto a Chell.  Before the patch it reported
<empty> as the touch_fw_version, and with this patch deployed it
showed a snippet of /var/log/messages that contained the touch FW
updater's logs as intended

Change-Id: I838a018ba90c739926ff3e654369e9e3c8a594d7
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/326152
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>