libbrillo: policy: Add GetUsbDetachableWhitelist for new USB policy
The newly added device policy setting |usb_detachable_whitelist| contains
a list of USB devices (identified by their USB VID:PID pairs) which can be
detached from the kernel drivers in order to use them in web application.
Add a GetUsbDetachableWhitelist method to retrieve this list of devices.
(as used by the permission_broker daemon)

The policy_all file was updated to include a usb_detachable_whitelist
policy with a couple of USB devices with VID:PID 0403:6001 and
413c:2105.

BUG=chrome-os-partner:50249
TEST=`FEATURES=test emerge-samus libbrillo permission_broker`

Change-Id: If4b06a85ab0b9641e615692449d434e41bd457e6