shill: Handle VPN reconnections on link down/up/changed
If the default physical service goes down or changes, and there is a
third party VPN connected at the time, put the VPN into the
Configuring state and notify the extension.  This gives the extension
an opportunity to reconnect and reconfigure the VPN tunnel when the
user roams to a different network.

This is an opt-in feature.  Existing VPN extensions don't support it, so
only extensions that set the new "reconnect" flag will receive the new
events.

Bug: None
BUG=chromium:514343
TEST=`FEATURES=test emerge-link shill`
TEST=install test VPN app and manually watch the JS console for events

Change-Id: I8916973e685314322ad0b7529968c2ed8ff9bd67