quipper: Remove CStringWithLength from PerfReader
This struct reflects the storage of string fields in raw perf data, but
it is not required for reading or storage. Instead, the strings can be
stored by themselves, and the storage size can be re-computed when it's
time to write strings back to raw data.

BUG=chromium:568870
TEST=unit tests pass

Change-Id: Id15ae13811b5099c6e3894c12244e24ae7923211
Reviewed-on: https://chromium-review.googlesource.com/329545
Commit-Ready: Simon Que <sque@chromium.org>
Tested-by: Simon Que <sque@chromium.org>
Reviewed-by: David Sharp <dhsharp@chromium.org>