power: Add support for "display off" mode with keyboard wakeup
This patch adds a new operating mode to powerd's WakeupController: in
addition to laptop, tablet and docked (external display with lid closed)
we now have "display off" (external display with lid open but display
brightness at 0). This mode is useful to use a Chromebook as a "smart
keyboard", in tablet configuration with the internal screen facing the
table.

It's natural to use the keyboard to wake up the device after idle
suspend in this mode. However, keyboard wakeup is gated by the EC which
tries to prevent "accidental" key presses from waking up the device in
tablet mode. Since the EC cannot know that the device has an external
display attached and is not used as a tablet (despite the lid-angle),
powerd has to disable this feature via MOTIONSENSE_CMD_KB_WAKE_ANGLE
(using a maximum wake angle of 360 effectively disables angle-based
wakeup inhibition completely).

BUG=chromium:552049
TEST=Played with Minnie in all the different configurations (lid angle
above or below 180, with or without external display, backlight on or
off) and confirmed that wakeup policy was as intended. Built with
FEATURES=test for Falco.

Change-Id: I64bafc0be9c83ccf4b4811deb0d7454de8e99775
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/329948
Reviewed-by: Dan Erat <derat@chromium.org>